requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/requirements.txt requirements/requirements.in
#
aiohttp-xmlrpc==0.4.1
aiohttp==2.2.3            # via aiohttp-xmlrpc, sir-bot-a-lot
apscheduler==3.3.1
async-timeout==1.2.1      # via aiohttp, sir-bot-a-lot
chardet==3.0.4            # via aiohttp, sir-bot-a-lot
distance==0.1.3
gidgethub==2.4.1
lxml==3.8.0               # via aiohttp-xmlrpc
mako==1.0.6               # via sir-bot-a-lot
markupsafe==1.0           # via mako, sir-bot-a-lot
multidict==3.1.0          # via aiohttp, sir-bot-a-lot, yarl
pluggy==0.4.0             # via sir-bot-a-lot
pytz==2017.2              # via apscheduler, tzlocal
pyyaml==3.12              # via sir-bot-a-lot
sir-bot-a-lot==0.1.1
six==1.10.0               # via apscheduler
slimurl==0.7.8            # via aiohttp-xmlrpc
tzlocal==1.4              # via apscheduler
uritemplate==3.0.0        # via gidgethub
yarl==0.11.0              # via aiohttp, sir-bot-a-lot
